requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
httpx
python-dotenv